add new endpoint(get user by id) to the rest controller and service

+++ src/main/java/com/example/realtimechatproject/repositories/UserRepository.java
@@ -10,7 +10,7 @@ import java.util.List;
 public interface UserRepository extends CrudRepository<UserEntity,Long> {
     UserEntity getFirstByHaslo(String haslo);
     List<UserEntity> getAllByHaslo(String haslo);
-    UserEntity getFirstById(Integer id);
+    UserEntity getFirstById(Long id);
 
     UserEntity getFirstBySurname(String surname);
     List<UserEntity> getAllBySurname(String surname);
